+++ src/pages/Admin.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+function Admin() {
+    return (
+        <div>
+            
+        </div>
+    )
+}
+
+export default Admin
